refactor: clean up the code

+++ src/main/resources/schema.sql
@@ -1,12 +1,12 @@
 CREATE TABLE formulas (
                           id SERIAL PRIMARY KEY,
                           formula_string VARCHAR(300) NOT NULL,
-                          x1 NUMERIC,
-                          x2 NUMERIC,
-                          x3 NUMERIC,
-                          x4 NUMERIC,
-                          x5 NUMERIC,
-                          result NUMERIC
+                          x1 BIGINT,
+                          x2 BIGINT,
+                          x3 BIGINT,
+                          x4 BIGINT,
+                          x5 BIGINT,
+                          result BIGINT
 );
 
 ALTER TABLE formulas ADD COLUMN created_at TIMESTAMP DEFAULT NOW();
